requirements.txt
absl-py==2.2.2
altair==5.5.0
asttokens==2.4.1
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.12.2
blinker==1.9.0
cachetools==5.5.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.2.0
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
flatbuffers==25.2.10
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
gTTS==2.3.2
h5py==3.13.0
idna==3.4
inexactsearch==1.0.2
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.9
narwhals==1.39.1
nest-asyncio==1.6.0
networkx==3.2
nltk==3.8.1
numpy==1.26.1
opt_einsum==3.4.0
optree==0.15.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.29.4
psutil==6.0.0
pure_eval==0.2.3
pyarrow==20.0.0
PyAudio==0.2.13
pydeck==0.9.1
Pygments==2.18.0
pyopencl==2024.2.7
pypiwin32==223
python-dateutil==2.9.0.post0
pytools==2024.1.14
pyttsx3==2.90
pytz==2025.2
pywin32==306
pyzmq==26.2.0
referencing==0.36.2
regex==2023.8.8
requests==2.31.0
rich==14.0.0
rpds-py==0.25.0
scikit-fuzzy==0.4.2
scikit-learn==1.6.1
scipy==1.11.3
silpa-common==0.3
six==1.16.0
smmap==5.0.2
soundex==1.1.3
soupsieve==2.5
SpeechRecognition==3.10.0
spellchecker==0.4
stack-data==0.6.3
streamlit==1.45.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
textblob==0.17.1
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.0.4
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
